internal/k8s: add buffered ResourceEventHandler
Add a buffered ResourceEventHandler. The buffer ensures that
notifications from various watchers are processed sequentially. That is
the underlying ResourceEventHandler never called concurrently.

Placing a lock around the underlying ResourceEventHandler raises the
question of the effect of that lock being held for an extended period of
time and thus the effect on the watchers TCP streams. The buffered
ResourceEventHandler includes a provision for a buffer as well as
notificaiton when that buffer is overrun and the sender, the watcher, is
blocked.

Signed-off-by: Dave Cheney <dave@cheney.net>