internal/contour: refactor internal caches
Requiring the cache.init() setup step is a grotty code smell and
prevented the translator having a valid zero value.

Replace the (frankyly) far to clever channel based cache with a good old
mutex. The caches will be simplified more in the future, probably
collapsing down onto a single cache of type.Any's, but for now this
is one step in unblocking the transformer being usable without an
initaliser.

Signed-off-by: Dave Cheney <dave@cheney.net>