cmd/contour: make oicd/gcp auth plugins build time options
This PR makes the oicd and gcp auth packages selectable at build time.
The intent is to avoid bundling these libraries in the contour image as
they are not needed when running `--incluster`, however they are needed
for local development when running `contour serve` locally.

The Makefile contains example syntax of how to use the `-tags` flag to
enable plugins as required.

Signed-off-by: Dave Cheney <dave@cheney.net>