internal/contour: Do not hash ClusterLoadAssignment's ClusterName (#187)
* internal/contour: Do not hash ClusterLoadAssignment's ClusterName

The ClusterLoadAssignment's ClusterName must match the corresponding
Cluster's ServiceName in the EDS config. Given that the ServiceName does
not have a character limit and is not being hashed, we should not hash
the ClusterName field in the ClusterLoadAssignment.

Signed-off-by: Alexander Brand <alexbrand09@gmail.com>

* internal/contour: add helper for computing service name.

The service name in the EDS config of a cluster must match the cluster
name of the corresponding ClusterLoadAssignments. Use the helper in both
places to avoid getting them out of sync.

Signed-off-by: Alexander Brand <alexbrand09@gmail.com>