deployment: optionally expose the Envoy admin inteface on an ClusterIP
This PR addresses the issue of debugging what configuration has been
pushed to Envoy. I (Dave) am at a point where I can look at an ingress
object and I will know how Envoy will react, but this level of spoon
bending is not available to others. For users of Contour they would
benefit from getting access to the Envoy admin interface so they can see
the resulting route, listerner, and cluster configs.

The previous documentation in the TROUBLESHOOTING document was less than
ideal as it encouraged users to edit the service connected to the
service load balancer, which would result in port 9001 binding to their
public IP.

This revised approach uses two service objects, the latter bound to a
ClusterIP, which offers a little more protection.

Signed-off-by: Dave Cheney <dave@cheney.net>