docs: recommend kubectl port-forward for envoy admin interface access
Rather than binding the admin interface to 0.0.0.0 or using a service
and kubectl proxy, recommend port forwarding directly to the admin
interface via the kubelet.

This has the advantages that can avoid the `--admin-address` flag our of
our example deployments, and as the admin interface remains bound to the
pod, it can't be accidentally exposed to other members of the cluster,
or the public internet.

Signed-off-by: Dave Cheney <dave@cheney.net>