internal/contour: properly flush long endpoint names from EDS cache
Updates #186

This PR fixes a condition whereby endpoints (clusterloadassignments)
could remain in the EDS cache after the endpoint has been removed from
the apiserver. This was partially fixed by #187, which dealt with the
case of accidentally hashing the servicename field of the
clusterloadassignment, however hashname was still being applied when
generating the key to remove the endpoint object from the EDS cache.

Apart from leaving unreferenced, as the endpoints' service would have
been removed also, entries in the EDS cache this bug was probably
benign so this is not critical to backport to 0.3.x.

Signed-off-by: Dave Cheney <dave@cheney.net>