internal/contour: stop CDS entries leaking in service update
Prior to this PR the translator treated an OnUpdate event for a service
as an Add. This could lead to conditions where Envoy clusters would
remain in the CDS cache if the service was edited.

This PR refactors service handling to _always_ model as an update,
making add and remove degenerate cases.

Signed-off-by: Dave Cheney <dave@cheney.net>