internal/contour: handle named service target ports correctly
Fixes #175

This PR fixes one of the longest running bugs in Contour, handling
services with named target ports. Curiously, it does so without using
the target port value.

The reason we can ignore the target port value is the endpoint
controller takes care of canonicalising named target ports to their
matching port numbers. This leaves us with three scenarios.

1. If the service's port is not named (targetport name/number is not
important), then the _must_ (the api server enforces this contract) be
only one port in the endpoint object. In this case CDS and EDS use the
port number of this single service as the CDS servicename.

2. If the service's port _is_ named, then that name will appear in the
endpoint object. In this case CDS and EDS will use the port's name as
the CDS servicename.

3. If the services' port is named, then there can be more than one port,
but they will all be named, and the names will appear in the endpoint
object as per point 2.

In all cases CDS and EDS can predict the servicename for the service
without having to reference each other explicitly.

This also allows EDS and CDS to operate seperately, whereas previously,
in attempting to fix this bug, a state machine created a one way
relationship between the service and endpoint objects.

With respect to the target port, if the target port is named then the
name can differ across matching pods. The endpoint controller will take
care of that by creating multiple endpoint subsets with a named port per
subset containing the endpoints that host entries listening on the
container port that matches the named port in that subset.

Signed-off-by: Dave Cheney <dave@cheney.net>