Add annotations to configure retries, max connections to a service
These are needed in order to scale a service up as well as help cope
with when a pod for a kubernetes service disappears unexpectedly. By
default Envoy has a limit of 1024 simultaneous connections to a cluster
and doesn't do any retries. A single http server backend of a service
can handle tens of thousands of requests per second however, and in
environments where that happens the low limit of envoy causes envoy to
circuit break those services rather than "flood" an unsuspecting backend
with requests.

By default, envoy doesn't retry requests, which means if a request gets
made to a pod which is the process of exiting / just stopped responding
to connections (Say it segfaulted) the request gets dropped and the
client sees an error rather than a retry. The retry annotations allow
changing this policy as appropriate in the environment

Signed-off-by: Cody Maloney <cody@emeraldcloudlab.com>