internal/e2e: add listener e2e test skeleton
This PR adds a scaffold for LDS tests. It uses the go-control-plane api
types to generate the configuration for the http connection manager. I
had initially decided not to use this method inside contour/listener.go
because of the barf worthy way of converting those types to a grpc
struct type.

For this e2e test I decided to use go-control-plane pkg's methods as a
way of double checking my hand rolled proto struct, I found a few small
errors which have been corrected in this PR, so it's good to double
check the data structures line up.

Signed-off-by: Dave Cheney <dave@cheney.net>