internal/envoy: use strict DNS for resolving the cluster
Fixes #228.

This enable us to define the contour cluster as a DNS name. I have confirmed
that `STRICT_DNS` also works when the address is an IP.

As mentioned by @davecheney before, with #158 and #228 fixed we should be able
to split Contour and Envoy into separate pods. Proof of concept:
https://gist.github.com/sevein/d30e1791fbc0db786884360486e70737.

Signed-off-by: Jesús García Crespo <jesus@sevein.com>