all: normalise logrus use
This PR finalises the use of logrus across the project. The various
TODO's related to logging can be taken care of in latter PRs.

The basic rules for using logrus in this project are

- A worker is reponsible for adding its own fields like `context` or
`resource`. Pass a `logrus.FieldLogger` to the worker's constructor and
let it add fields as necessary.
- The name of the `logrus.FieldLogger` or `logrus.StdLogger` should be
`log` in formal parameters and local variables.
- If the function is not expected to add more fields, the type of it's
logger should be `logrus.StdLogger`.
- `logrus.FieldLogger` / `logrus.StdLogger` should be embedded by value
into long lived types that need to log. They should apply necessary
fields during construction.

Signed-off-by: Dave Cheney <dave@cheney.net>