internal/contour: derive cluster.service_name from targetport
Fixes #243

Hopefully this is the _last_ issue relating to the interaction between
CDS, EDS, and the endpoint object.

The logic now follows that _iff_ a service port is named, use that name
for the cluster.service_name field in CDS. If the service port is
unnamed then derive the name from the service port's targetport, and use
that in cluster.service_name.

Signed-off-by: Dave Cheney <dave@cheney.net>