internal/contour: generate stable service names for unnamed service ports
Fixes #247

Fingers crossed this finally closes the last issue with the EDS and CDS
servicename's not matching.

On the CDS side, the cluster.service_name field is always defined from
the namespace/name of the service document, and, if provided a slash and
the name of the service port. If this is blank, then the slash is also
elided. The end result is in the case that there is a single serviceport
in the service spec, the service_name generated for that CDS entry will
be simple "namespace/name". In all other cases, the api server requires
that service ports be named, so the CDS's service_name entry will be
"namespace/name/serviceportname".

On the EDS side, we only have the contents written by the endpoint
controller, but that is enough. If the service's port was named in the
service document, then that name will be present in the endpoint
document. If not, then the name field will be blank, which generates a
corresponding service name of "namespace/name", matching the logic
described above in CDS.

Signed-off-by: Dave Cheney <dave@cheney.net>