internal/contour: 301 upgrade routes on a per ingress basis
Fixes #250

Prior this PR if _any_ ingress which mentioned a vhost used the
ingress.kubernetes.io/force-ssl-redirect: "true" annotation, then all
routes would be 301 upgraded via the route.virtualhost.requireTls flag.

This PR changes the behaviour so per ingress (that is all routes in an
ingress object), not per vhost, the annotation is applied and this
causes the route.RouteRoute action to be replaced with a
route.Route_Redirect action. This has the same net effect as the
requireTls virtualhost flag, but is applied per route.

Signed-off-by: Dave Cheney <dave@cheney.net>