integration: add generator and service fixture
Updates #291

To stress test contour's gRPC limits we need to generate a bunch of mock
services. The services do not need to have any matching pods (although
that is an improvement to be landed in the future).

This commit ads a basic generator and sample service data which can be
loaded into a throwaway namespace to test envoy and contour's limits.

Signed-off-by: Dave Cheney <dave@cheney.net>