internal/grpc: filter resource values from DiscoveryRequest
Construct a filter function from the DiscoveryRequest.ResourceNames and
pass it to the cache's Values method.

If the set of returned values is empty, do not reply to the stream
response. This has the effect that when the cache is notified, resources
which do not match the current filter, from the current discover
request, will return an empty set of values, so we can skip sending the
response to Envoy.

Signed-off-by: Dave Cheney <dave@cheney.net>