internal/contour: reuse discoveryrequest if nothing matches filter
Fixes #363

This PR fixes a bug whereby stream watchers would jam if nothing matched
the supplied filter. This bug affected all xDS watchers, but would be
apparent in LDS.

The underlying cause was the sequence of operations in xDS is

1. receive request from envoy
2. look up the resource in the map of known types
3. construct a filter from the request and wait for the resource to
change
4. once we get a notification, filter the response
4a. if there are no resources after filtering then go to 1.
5. send response and go to 1.

The bug was 4a caused the stream watcher to block waiting for another
request object, but Envoy will not send one because we haven't replied
to the previous. The fix is in 4a to jump to 3, thereby reusing the
request that is outstanding.

Lamentably there is no test because the e2e tests use fetch, not stream.
This is super sucky and I will raise a bug to address this, but for the
moment this PR has been manually verified to address the issue.

Signed-off-by: Dave Cheney <dave@cheney.net>