design: note that ingressroute is multi-team, not multi-tenant
Update ingressroute spec to acknowledge that the design is multi-team,
not multi-tenant. In doing so this resolves the security issue noted
previously.

We define multi-tenant to be a cluster who's members are not known or
trusted by each other, whereas multi-team share a level of trust, or
shared organisational responsibility. In effect the security of
multi-team clusters is delegated to the organisations reporting
structure, not enforced in code.

Signed-off-by: Dave Cheney <dave@cheney.net>