internal/contour: introduce dedicated EndpointsTranslator
Inside Contour Servics, Ingress, and Secret objects are interelated,
however Endpoints are both orthogonal to the operation of the
Translator, and follow a different update regieme. While the former
three are updated infrequently, by humans, the latter is updated
frequently by a machine.

To reflect this difference, and make refactoring on the Translator
easier in the future, introduce a new type, EndpointsTranslator. The
EndpointsTranslator implements cache.ResourceEventHandler, but responds
only to *v1.Endpoints objects (the corresponding role has been removed
from contour.Translator). This makes it the sole place where Endpoint
updates are processed from the API server, and cached for Envoy.

To reflect this the EndpointsTranslator is woven through cmd/contour
independant of the k8s.Buffer, and glued into the GRPC server when
constructing the Stream handler.

Signed-off-by: Dave Cheney <dave@cheney.net>