deployment: switch the order of envoy and contour containers
Updates #420

Prior to this PR the envoy container appeared first in the
deployment.spec so would likely start before the contour container. This
would cause envoy to fail to be able to contact contour immediately upon
startup and would then fall back to a retry behaviour.

With this change, contour appears first in the spec so will likely start
before contour, which agrevates a failure mode reported in issue #420.

Signed-off-by: Dave Cheney <dave@cheney.net>