internal/grpc: always send discoveryresponse with zero results
Fixes #420

In #350 I added the ability to filter the results of a discovery
request to address the scalability issues in sending the full copy of
the xDS caches to envoy on every request -- this was especially imporant
for EDS which would send an entire copy of the EDS cache to each
StreamEndpoints listener for every change in k8s endpoints.

As part of this change I recognised that there would be situations where
a notification that the cache had changed would result a wakeup and
calls to values for events which didn't match the filter. At the time I
mistakenly thought that if a wakeup was triggered that didn't match the
filter this would cause zero length responses to be sent to Envoy
causing EDS entries (and others) to pop in and out of existance.

For example, if you were watching for endpoints matching "A", then an
update happened to endpoints matching "B", I believed that when the "A"
watcher ran it would receive an empty set of endpoints and send those
back to envoy, effectively erasing the active set on the Envoy side.

In fact what happens, and what still happens, is a notification that the
cache to the EDS cache wakes up all watchers, they run Values(), passing
in their filter. This returns whatever results match the filter, which
critically will be whatever is in the eds cache at the time. If there
are several results for "A" and the EDS cache was updated for "B", then
both "A" and "B"'s watchers will be woken up, they'll both call Values()
and then send the set of endpoints back to Envoy. In that case the
change to B will be reflected and the change to A will be a no-op.

What I tried to avoid was the update to "B" causing the Values for "A"
running and finding no endpoints, because I misunderstood that an update
to "B"'s endpoints would make it appears that "A"'s endpoint set was
empty. This was not the case, and will only be the case if "A"'s
endpoint set is actually empty.

In the case that "A"'s endpoint set was empty, the logic prior to
this PR would skip the notification and wait for the number of endpoints
to grow above zero. This _sort of_ made sense when CDS and EDS entries
would arrive after startup, but was fundamentally flawed. EDS entries
with zero values are a valid response and don't represent a transient
state of a filter not matching.

Where this bug arose is during startup Envoy _may_ query (depends on
timing) EDS values synchronously or asynchronously, and in the former
Envoy will first look in CDS for the names of all clusters, then query
EDS for the endpoints of all the clusters previously mentioned. If the
EDS entry had zero endpoints, the notification would be suppressed and
thus Envoy's startup would stall.

The fix to this bug is simple, always send the response we get back from
Values. If it's empty, then that is because there are no endpoints
registered, regardless of the filter applied to trim out endpoints that
were not requested.

Again, as with #350 there is no test, other than the existing test which
prove that existing functionality (so far as it is covered) is
unaffected. However manual testing, in the form of moving contour first
in the deployment descriptor which triggers Envoys synchronous startup
is part of #422, which should give some coverage.

Signed-off-by: Dave Cheney <dave@cheney.net>