internal/dag: canonicalise service port during Recompute
Updates #211
Fixes #462

Handling the IntStr nature of Ingress's Backend.SevicePort field has
been a thorn in Contour's side for a long time (see #211). This PR
resolves both #211 and #462 by resolving the port number during DAG
recomputation if a named port is in use.

We store the selected port number on dag.Service and construct a tupple
of namespace/name/port to treat each service separately, ie. a service
with port 9000 is different to the port 9001 on the same service object.

In doing so we remove the requirement to store the particular
v1beta1.ServiceBackend on the Route object, which resolves the issue of
an ingressroutev1 object lacking that object. The route object doesn't
need to hold this information anyway, it's already present on the
Service object as described above.

Signed-off-by: Dave Cheney <dave@cheney.net>