internal/{grpc,e2e}: use stream in testing
Fixes #373

This PR fixes a bug in the grpc stream helper whereby streaming would
stall if the cache was empty when the stream request arrived. The stream
would be unblocked as soon as an update landed to the cache. In theory
this could have blocked startup of Envoy if, say, there were no ingress
records present in the API server.

To ensure this is tested the e2e tests have been updated to issue a
single stream req/resp pair, rather than using the fetch mechanisms,
this gives us some high level integration testing of the grpc streaming
mechanism.

Signed-off-by: Dave Cheney <dave@cheney.net>