internal/dag: add support for ingressroute delegation
IngressRoute (IR) resources can either point to backend services, or delegate
a path to another IR. While building the DAG out of the IRs and routes
within each, we determine whether the route delegates to another IR.
When it does, we find the referenced IR and process it accordingly.

This processing is applied recursively until we traverse all the edges
and find a leaf route, or a route that does not delegate to another IR.
This is the base case, in which we grab the services, and add them to
the routes of the corresponding VHost.

Signed-off-by: Alexander Brand <alexbrand09@gmail.com>