internal/dag: introduce SecureVirtualHost
Fixes #476

This PR introduces a second kind of root vertex, a SecureVirtualHost.
Unlike the existing VirtualHost, this type represents HTTPS vhosts.

This change resolves a number of ambiguities in the DAG

- a TLS backed vhost is no longer indicated by the presence of at least
one child secret. Now the presents of a secret only exists on a
SecureVirtualHost -- Secrets are no longer children of VirtualHosts.
- this prevents the intermediate situation where a TLS backed site would
appear on port 80 because the secret was not present. Now if the secret
is not present the LDS walker can skip it entirely.
- This gives a place to hang TLS parameters like protocol version from,
fixing #476.

Signed-off-by: Dave Cheney <dave@cheney.net>