internal/e2e: fix failing force-redirect tests
This is a case where I think the previous translator got this wrong. In
the case where a vhost is mentioned on two ingresses, if one of those
ingresses uses TLS, then the routes for both should be available on both
http and https. Then if one of those ingresses uses the force-redirect
annotation, that annotation only applies to the routes on that ingress,
leaving the routes defined on the other ingress not 301 upgraded.

Confusing? Yes. This is INGRESS!

Signed-off-by: Dave Cheney <dave@cheney.net>