internal/*: detect the cycle before following the problematic edge
When a cycle is detected, the ingressroute that produces the cycle is
the one that is considered invalid. Before this change, we were visiting
the target ingressroute and making that one invalid.

For example, assuming the following ingressroute delegation is setup:

A => B => A

IngressRoute B is considered invalid because it produces the cycle,
instead of IngressRoute A.

Signed-off-by: Alexander Brand <alexbrand09@gmail.com>