internal/dag: generate secure vhosts first
Fixes #484

If a vhost is mentioned in two or more ingress objects, and one of those
objects includes a TLS stanza then depending on the order these
ingresses are processed during DAG.recompute -- which is not defined --
then routes may appear, or not appear, on secure vhosts depending on if
the secure vhost was defined at the point the route from the other
ingress was being inserted.

Solve this by making two passes of the ingress document list. The first
pass sets up the secure vhosts and secrets, the second will apply routes
to their secure counterparts if they exist.

internal/route will take care of pruning ingresses that have no routes.

TODO: if a vhost with tls is mentioned more than once the results will
be undefined. Blame the ingress spec for this.

Signed-off-by: Dave Cheney <dave@cheney.net>