internal/cluster: restore previous servicename calculation
EDS and CDS communicate via the service name, not the cluster.name. This
is important because EDS computes the name of the CDS's cluster's
service name in the blind using some logic encoded in the servicename
helper.

This logic was accidentally discarded in a previous commit in this PR
because I had forgotten the magic.

Signed-off-by: Dave Cheney <dave@cheney.net>