internal/contour: sort weighted upstream list
The slice of weighted upstreams is generated by iterating over a map so
is not stable. Envoy doesn't care about this, but our test fixtures do,
so to prevent flakes we sort the list of upstreams by name.

Protip: If you're running Go 1.10 or later, use go test -count=1 to
disable test caching if you think you have a test flake.

Signed-off-by: Dave Cheney <dave@cheney.net>