internal/dag: fix IngressRouteStatuses.version computation
Updates #445

Cleanup the computation of dag.version, and fix a bug where the version
assigned to the IngressRouteStatus was not incremented, thus
representing the version counter for the dag that would be replaced by
recompute.

The new value of 2 in the test fixture is caused by dh.Insert() calling
recompute internallary, then dh.Recompute() is called directly to
retrieve the current IngressRouteStatus.

Signed-off-by: Dave Cheney <dave@cheney.net>