internal/dag: combine ingressroute status with dag roots
Updates #445

This PR merges the return values from DAG.recompute; the dag.dag and the
[]Status list, into one, the dag.dag. Thus a dad.dad's set of roots, and
the status while generating those roots are bound together.

This has the side effect of removing the return parameter from
DAGAdater's OnAdd/Update/Remove methods, allowing it to fulfil the
cache.ResourceEventHandler interface directly.

This PR also merges the dag.dag.version and IngressrouteStatus.version
field as they were derrive from the same place, so storing the value
twice had no value. The IngressrouteStatus.GetVersion accessor was not
used and was removed. This meant that IngressrouteStatus was just a
wrapper around []Status, so it was also removed.

A note about skew between the contents of dag.dag.roots and
dad.dag.status:

Previously d.Recompute returned the set of status objects which was
passed to d.setIngressRouteStatus, then the udpateListener, Route, and
Cluster methods were called. Now d.Recompute places the status on the
d.current field, and d.setIngressRoute retrieves it via the DAG.Statuses
method.

This PR retains the observed behaviour; status is updated before the
gRPC caches. Neither before or after this PR is there a guarentee that
the gRPC caches would be updated with the _same_ version of the dag that
d.setIngressResourceStatus is uses. It was, and still is possible that
each of the four update methods, d.setIngressRouteStatus,
d.updateListeners, d.updateRoutes, and d.updateClusters could use a
different version of the dag, either via DAG.Visit or DAG.Statuses if
multiple k8s watchers fire at the same time.

Signed-off-by: Dave Cheney <dave@cheney.net>