internal/dag: make dag.Compute idempotent
Updates #445

This is an early PR of the infrastructure to address the O(N) cost of
DAG.Recompute.

This PR replaces DAG.Recompute with a new method, DAG.Compute which
returns a interface with Visit and Statuses methods. This means the
current dag.dag is no longer stored on the parent dag.DAG value.

The result of this is computation of a dag.dag from the source dag.DAG
(I really have to find better names) is now idempotent. The result from
calling DAG.Compute is stable, you can visit it as many times as you
want and it will not change from under you, nor will the values reported
by Statuses skew with respect to Visit.

The intention of this change is to decouple the following

1. DAG.Insert, Remove operations arrising from k8s watchers firing.
These are always handled syncronously inside the watcher loop.
2. DAG.Compute, is called synchronously inside the same loop as 1, but
can in the future be handed off on a channel to another process. That
is, passing the dag.DAG to something else which will call it's Compute
method at some later time.
3. The resulting value from calling DAG.Compute can be handed off again,
or processed in parallel as it is now stable.

One side effect of this change is calles to the debug handler call
DAG.Compute, rather than visiting the most recently computed dag. This
shouldn't be a problem in operation and we can add a cache if this shows
up in profiles.

Signed-off-by: Dave Cheney <dave@cheney.net>