internal/contour: split and rename DAGAdapter
Updates #445

This PR renames and splits DAGAdapter into two new types;
ResourceEventHandler, which continues to adapt the k8s watcher interface
of the same name to the dag.Builder Insert and Remove API, and a
CacheHandler which is responsible for regenerating the contents of the
various xDS caches based on the results of calling dag.Builder.Compute.

In this PR CacheHandler remains embedded inside ResourceEventHandler to
keep the change surface small, but a subsiquent PR will place the former
behind an interface making the objects independant.

Signed-off-by: Dave Cheney <dave@cheney.net>