internal/metrics: automatically register metrics during NewMetrics
Automatically register metrics during NewMetrics now we no longer need
to choose if we need to register the GoCollector during tests.

Also return a pointer to the metrics.Metrics value, not a value. Even
though this doens't matter right now as metrics.Metrics contains a map,
which is itself a pointer, in the future this will change. This also
showed a bug where metrics were not being registered during e2e tests
causing them to be thrown away in SetIngressRouteMetrics because the map
inside the zero value of a metrics.Metrics was empty.

Signed-off-by: Dave Cheney <dave@cheney.net>