internal/contour: add HoldoffNotifier
Fixes #577

This PR introduces a HoldoffNotifier implementation that delays the
notification of the underlying Notifier by 100 ms since the previous
call to HoldoffNotifier.OnChange or 500ms, whichever is the smaller.

The 100ms/500ms values are currently hard coded, a future PR may make
them configurable, which will allow this type to be used in the e2e
tests -- currently the e2e tests go directly to the
contour.CacheHandler Notifier.

Signed-off-by: Dave Cheney <dave@cheney.net>