internal/contour: ignore endpoint updates if both old and new have zero endpoints
Before this change, we were ignoring endpoint update events that had
zero endpoints in the updated object. This was a problem because we were
not handling the case where we went from N endpoints to zero endpoints.

With this change, we only ignore the update if we are going from zero
endpoints to zero endpoints. Otherwise, we handle the update like any
other.

Signed-off-by: Alexander Brand <alexbrand09@gmail.com>