internal/dag: use vararg for vhost/svhost aliases
Updates #451

Move the aliases parameter for vhost/svhost to the end of the formal
parameters and make it variadic. This means the default invocation can
avoid passing []string{} as a dummy parameter. Also, an empty vararg is
nil so we can remove a whole bunch of dummy parameters from the builder
test fixtures.

Signed-off-by: Dave Cheney <dave@cheney.net>