internal/dag: move svhost alias assignment outside lookupSecureVirtualHost
Updates #625

Currently lookupSecureVirtualHost takes a slice of aliases which is used when
lazy initalising the svhost record returned to the caller. This is a
problematic API for several reasons.

1. the aliases slice is passed around to all lookupSecureVirtualHost
call sites, this is redundant as aliases is not consulted if the svhost exists in the cache.
2. as aliases are not consulted if the svhost is cache, if a different
set of aliases is provided it is not clear to the caller that the
alias details will not be updated. Assuming it is known at the call site
that the svhost has been lookup up previous, nil can be passed for the
aliases set without penalty.

To resolve this situation we move responsibility for assigning the
aliases slice outside the call to lookupSecureVirtualHost. We also take
on the responsibility of making sure lookupSecureVirtualHost is called
once, before any callers of the same function with a view to adding
routes.

In general the use of the aliases parameter highlights a general pattern
in the dag builder; we either lookup a service or vhost with a view to
configuring its properties once, or we lookup with a view to adding a
child. We never do both in the same place, and the former is intended to
be an idempotent action, while the latter expects that the former has
already taken place.

As the same logic applies to lookupVirtualHost so it is included in this
change.

Signed-off-by: Dave Cheney <dave@cheney.net>