internal/dag: set all delegate routes orphaned initially
Address TODO from #629

The arguments to builder.setOrphaned already identify an ingressroute
uniquely by name and namespace, so there is no need to check if we've
already set the orphaned status bit as we will only ever encounter this
ingressroute once.

Also, as the orphaned bit is only set in a single pass over the set
of ingressroutes we don't need to check if we're setting a bit that was
previously unset -- the logic is simply, if you're a delegate
ingressroute, you are orphaned until proven otherwise.

Lastly, pass the ingressroute, not its name and namespace to
setOrphaned, this makes it harder to transpose the name and namespace
parameters (not that i've done that, several times).

Signed-off-by: Dave Cheney <dave@cheney.net>