internal/dag: move service weight assignment to lookupService
Updates #581

Move service weight assignment from Route.addService to
builder.lookupService. This means services with different route
weightings are considered different from the POV of the DAG -- each call
to lookupService with a different weight will result in a new service
entry in the builder's cache.

However, this does not affect the name of the CDS entry as service
weights are a property of RDS, not CDS.

Signed-off-by: Dave Cheney <dave@cheney.net>