internal/contour: update route vistor for cluster name change
Updates #581

Update the RDS route visitor for the previous change to CDS
cluster name field.

Rearrange the helpers such that servicename generates the CDS cluster
name for use by EDS; the name of the field is servicename. This frees
clustername for use by the RDS and CDS visitors for generating the
canonical name of the cluster.

Signed-off-by: Dave Cheney <dave@cheney.net>