append x-request-start header to incoming HTTP requests
Some tracing applications support reporting on queuing time for incoming
requests. Generally this is implemented by adding a `X-Request-Start`
timestamp header to the requests as early as possible in the request
path.

New Relic [1] and datadog [2] are two examples of tools that support
this header.

The envoy docs [3] say:

> Custom request/response headers can be added to a request/response at
> the weighted cluster, route, virtual host, and/or global route
> configuration level

I've opted to add them at the cluster level, based on a slack discussion
with Steve Sloka [4]. I'm happy to change that if requested though - my
envoy knowledge is pretty low, and I'm not sure what the tradeoffs of
each location are.

The header is added to all requests unconditionally, as suggested in the
original github issue.

Fixes #646

[1] https://docs.newrelic.com/docs/apm/applications-menu/features/request-queue-server-configuration-examples
[2] https://github.com/DataDog/dd-trace-rb/blob/master/docs/GettingStarted.md#http-request-queuing
[3] https://www.envoyproxy.io/docs/envoy/latest/configuration/http_conn_man/headers#custom-request-response-headers
[4] https://kubernetes.slack.com/archives/C8XRH2R4J/p1538487259000100

Signed-off-by: James Healy <james@yob.id.au>