enable automatic gzip compression of responses
This unconditionally enables envoys gzip http_filter[1], which will
compress responses if the request indicates it can handle it (via a
"accept-encoding: gzip" request header).

I've enabled it with no options, which means it gets the defaults, as
listed at [2]. In envoy 1.7.0, they're:

    {
      "memory_level": 5,
      "content_length": 30",
      "compression_level": "DEFAULT",
      "compression_strategy": "DEFAULT",
      "content_type": [“application/javascript”, “application/json”, “application/xhtml+xml”, “image/svg+xml”, “text/css”, “text/html”, “text/plain”, “text/xml”],
      "disable_on_etag_header": false,
      "remove_accept_encoding_header": false,
      "window_bits": 12
    }

Note the list of content types that will be compressed. It's possible
some users will want to customise this list, but for now I haven't
exposed that option.

Fixes #310

[1] https://www.envoyproxy.io/docs/envoy/latest/configuration/http_filters/gzip_filter
[2] https://www.envoyproxy.io/docs/envoy/latest/api-v2/config/filter/http/gzip/v2/gzip.proto

Signed-off-by: James Healy <james@yob.id.au>