Makefile: add errcheck target
Add an errcheck target for running kiselk's errcheck tool.

This check is not enabled in make check because it is too noisy; there
are legitimate cases to ignore an error and I'm not going to add helpers
to make errcheck happy, the value gained is too low.

However, for a few of the points that errcheck raises I have addressed
them by refactoring the code; this improved the cleanup handling in
internal/e2e, so win/win I say!

Signed-off-by: Dave Cheney <dave@cheney.net>