internal/e2e: remove context.WithTimeout from steam* helpers
Updates #705

The 100ms timeout was too short in some circumstances; maybe there are
clock skew issues on the CI host.

Rather than playing the "let's increase the timeout and hope" game,
remove the timeout. There are no circumstances where the e2e tests do
not expect a result so if one goes missing this is a serious bug.

Signed-off-by: Dave Cheney <dave@cheney.net>